Extract a generalised SuspendableService from SuspendableTimerService.
