+++ ee-feature-pack/galleon-shared/src/main/resources/modules/system/layers/base/org/jboss/as/clustering/common/main/module.xml
@@ -31,6 +31,7 @@
         <module name="org.wildfly.clustering.context"/>
         <module name="org.wildfly.clustering.function"/>
         <module name="org.wildfly.clustering.service"/>
+        <module name="org.wildfly.clustering.server.spi"/>
         <module name="org.wildfly.common"/>
         <module name="org.wildfly.security.elytron-private"/>
         <module name="org.wildfly.service"/>
